[FIX]-Ajuste de forma farmaceutica

+++ base_solution/models/sale_order.py
@@ -10,23 +10,28 @@ class SaleOrder(models.Model):
     def action_confirm(self):
         raw = []
         values = []
-        for line in self.raw_material:
-            dic = {
-                'product_id': line.product_id.id,
-                'product_qty': line.product_qty,
+        validation = len(self.order_line)
+        
+        if validation == 1:
+            for line in self.raw_material:
+                dic = {
+                    'product_id': line.product_id.id,
+                    'product_qty': line.product_qty,
+                }
+                raw.append((0,0,dic))
+            values = {
+                'product_tmpl_id': self.order_line.product_id.product_tmpl_id.id,
+                'bom_line_ids': raw,
             }
-            raw.append((0,0,dic))
-        values = {
-            'product_tmpl_id': self.order_line.product_id.product_tmpl_id.id,
-            'code': '001',
-            'bom_line_ids': raw,
-        }
-        record_ids = self.env['mrp.bom'].search([('product_tmpl_id', '=', self.order_line.product_id.product_tmpl_id.id)])
-        if self.order_line.product_id.name == 'Generico cotizador':
-            record = self.env['mrp.bom'].search([('product_tmpl_id', '=', self.order_line.product_id.product_tmpl_id.id)]).bom_line_ids.bom_id
-            delete = self.env['mrp.bom.line'].search([('bom_id', '=', record._origin.id)])
-            delete.unlink()
-            for record in record_ids:
-                record.write(values)
-        res = super(SaleOrder, self).action_confirm()
+            record_ids = self.env['mrp.bom'].search([('product_tmpl_id', '=', self.order_line.product_id.product_tmpl_id.id)])
+
+            if self.order_line.product_id.name == 'Generico cotizador':
+                record = self.env['mrp.bom'].search([('product_tmpl_id', '=', self.order_line.product_id.product_tmpl_id.id)]).bom_line_ids.bom_id
+                delete = self.env['mrp.bom.line'].search([('bom_id', '=', record._origin.id)])
+                delete.unlink()
+                for record in record_ids:
+                    record.write(values)
+                res = super(SaleOrder, self).action_confirm()
+        else:
+            res = super(SaleOrder, self).action_confirm()
         return res
